Latest Stable

+++ Client/wwwroot/js/Dashboard.js
@@ -1,15 +1,48 @@
 ﻿$(document).ready(function () {
     $('#TabelMasterData').DataTable({
-        dom: "<'row justify-content-between align-items-center'<'col-md-3'l><'d-flex align-items-center justify-content-center'<'mr-3'B><'mt-1'f>>>" +
+        dom: "<'row'<'col-md-3'l><'col-md-5'B><'col-md-4'f>>" +
             "<'row'<'col-md-12'tr>>" +
             "<'row'<'col-md-5'i><'col-md-7'p>>",
+        columnDefs: [{
+            targets: -1,
+            className: 'ignore'
+        }],
         buttons: [
-            'copy', 'csv', 'excel', 'pdf', 'print'
+            {
+                extend: 'copy',
+                exportOptions: {
+                    columns: ':not(.ignore)'
+                }
+            },
+            {
+                extend: 'csv',
+                exportOptions: {
+                    columns: ':not(.ignore)'
+                }
+            },
+            {
+                extend: 'excel',
+                exportOptions: {
+                    columns: ':not(.ignore)'
+                }
+            },
+            {
+                extend: 'pdf',
+                exportOptions: {
+                    columns: ':not(.ignore)'
+                }
+            },
+            {
+                extend: 'print',
+                exportOptions: {
+                    columns: ':not(.ignore)'
+                }
+            }
         ],
         "filter": true,
         "orderMulti": false,
         "ajax": {
-            "url": "https://localhost:44327/api/accounts/masterdata",
+            "url": "accounts/masterdata",
             "datatype": "json",
             "dataSrc": "result"  
         },
@@ -21,12 +54,12 @@
                 "data": "fullName"
             },
             {
-                "data": "gender",
+                "data": "gender"
             },
             {
                 "data": "birthDate",
                 "render": function (data, type, row) {
-                    return row["birthDate"].toString().substring(0,10);
+                    return moment(row["birthDate"]).format('LL');
                 }
             },
             {
@@ -53,60 +86,50 @@
                 "data": null,
                 "orderable": false,
                 render: function (data, type, row, meta) {
-                    return `<button class="btn btn-primary" onClick="GetUpdateModal('${row.nik}')" data-toggle="modal" data-target="#updateForm">Edit</button><br>
-                            <button class="btn btn-primary" onClick="Delete('${row.nik}')">Delete</button>`;
+                    return `<div class="row">
+                                <div class="col-md-5">
+                                <button class="btn btn-warning fa-solid fa-pen-to-square" onClick="GetUpdateModal('${row.nik}')" data-toggle="modal" data-hover="tooltip"  data-placement="right" title="Edit" data-target="#updateForm">
+                                </div>
+                                <div class="col-md-5">
+                                </button> <button class="btn btn-danger fa-solid fa-trash-can" onClick="Delete('${row.nik}')" data-hover="tooltip" data-placement="right" title="Delete"></button>
+                                </div>
+                            </div>`;
                 },
+                "autowidth": true
             },
         ],
     });
 });
 
-
-$.ajax({
-    url: "https://localhost:44327/api/accounts/masterdata",
-    success: function (results) {
-        var result = results.result;
-        let genderDataArray = [0, 0];
-        $.each(result, function (key, val) {
-            if (val.gender == "Male") {
-                genderDataArray[0] += 1;
-            }
-            else {
-                genderDataArray[1] += 1;
-            }
-        });
-        console.log(genderDataArray);
-        var options = {
-            chart: {
-                type: "pie"
-            },
-            series: genderDataArray,
-            labels: ["Male", "Female"]
-        }
-        var chart = new ApexCharts(document.querySelector("#genderchart"), options)
-        chart.render();
-    }
-})
-
 let universityData = [];
 $.ajax({
-    url: "https://localhost:44327/api/universities/",
+    url: "universities/getall",
     success: function (results) {
-        let result = results.result;
+        var result = results.result;
+        console.log(result);
         let text = `<option selected value="">Choose...</option>`;
         $.each(result, function (key, val) {
             universityData.push({ x: val.name, y: 0 });
-            text += `<option type="number" value="${val.id}">${val.name}</option>`
+            text += `<option type="number" value=${val.id}>${val.name}</option>`
         });
         $("#uni").html(text);
     }
 })
 
 $.ajax({
-    url: "https://localhost:44327/api/accounts/masterdata",
+    url: "accounts/masterdata",
     success: function (results) {
         var result = results.result;
+        let genderDataArray = [0, 0];
         $.each(result, function (key, val) {
+
+            if (val.gender == "Male") {
+                genderDataArray[0] += 1;
+            }
+            else {
+                genderDataArray[1] += 1;
+            }
+
             for (var i = 0; i < universityData.length; i++) {
                 if (universityData[i].x == val.university_Name) {
                     universityData[i].y += 1;
@@ -114,12 +137,35 @@ $.ajax({
                 }
             }
         });
+
         var options = {
             chart: {
-                height: 417,
+                type: "pie",
+                selection: {
+                    enabled: true
+                },
+                toolbar: {
+                    show: true,
+                    tools: {
+                        download: true,
+                    },
+                    offsetX: -10,
+                    offsetY: 0
+                }
+            },
+            series: genderDataArray,
+            labels: ["Male", "Female"]
+        }
+        var chart = new ApexCharts(document.querySelector("#genderchart"), options)
+        chart.render();
+
+        var options = {
+            chart: {
+                height: 420,
                 type: "bar"
             },
             series: [{
+                name: "counts",
                 data: universityData
             }],
             xaxis: {
@@ -142,32 +188,42 @@ function Insert() {
         Password : $("#password").val(),
         Degree : $("#degree").val(),
         GPA : $("#gpa").val(),
-        University_Id : $("#uni").val(),
+        University_Id: $("#uni").val(),
         Salary : $("#salary").val(),
         Birthdate: $("#birthdate").val()
     };
     $.ajax({
-        url: "https://localhost:44327/api/accounts/register",
+        url: "accounts/register",
         type: "POST",
-        headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json',
-        },
         dataType: 'json',
-        data: JSON.stringify(emp),
+        data: emp
     }).done((result) => {
-        alert("Registration success");
-        location.reload();
+        Swal.fire({
+            position: 'center',
+            icon: 'success',
+            title: 'New employee data has been saved',
+            showConfirmButton: false,
+            timer: 1500
+        }).then(function () {
+            location.reload();
+        }
+        );
     }).fail((error) => {
-        alert("Registration error");
-})
+        Swal.fire({
+            icon: 'error',
+            title: 'Registration failed',
+            text: "Invalid data",
+            footer: '<a href="">Check the backend programming</a>'
+        }).then(function () {
+        }
+        );
+    })
 }
 
 function GetUpdateModal(nik) {
     $.ajax({
-        url: `https://localhost:44327/api/employees/${nik}`,
-        success: function (results) {
-            var result = results.result;
+        url: `employees/get/${nik}`,
+        success: function (result) {
             $('#unik').attr('value', `${nik}`);
             $('#ufirstname').attr('value', `${result.firstName}`);
             $('#ulastname').attr('value', `${result.lastName}`);
@@ -175,19 +231,13 @@ function GetUpdateModal(nik) {
             $('#uphone').attr('value', `${result.phone}`);
             $('#ubirthdate').attr('value', `${result.birthDate}`.toString().substring(0, 10));
             $('#usalary').attr('value', `${result.salary}`);
-            var currentgender; var altgender; var altgenderval;
+            //var currentgender; var altgender; var altgenderval;
             if (result.gender == 0) {
-                currentgender = "Male";
-                altgender = "Female";
-                altgenderval = 1;
+                $('#ugender').val("0")
             }
             else {
-                currentgender = "Female";
-                altgender = "Male";
-                altgenderval = 0;
+                $('#ugender').val("1")
             };
-            $('#ugender').html(`<option selected>${currentgender}</option>
-                                <option type="number" value="${altgenderval}">${altgender}</option>`);
         }
     });
 }
@@ -205,32 +255,63 @@ function Update() {
         birthDate: $("#ubirthdate").val()
     };
     $.ajax({
-        url: "https://localhost:44327/api/employees",
+        url: `employees`,
         type: "PUT",
-        headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json',
-        },
         dataType: 'json',
-        data: JSON.stringify(emp),
+        data: emp
     }).done((result) => {
-        alert("Edit success");
-        location.reload();
+        Swal.fire({
+            position: 'center',
+            icon: 'success',
+            title: `The new data of ${emp.nik} has been saved`,
+            showConfirmButton: false,
+            timer: 1500
+        }).then(function () {
+            location.reload();
+        }
+        );
     }).fail((error) => {
-        alert("Edit error");
-        location.reload();
+        Swal.fire({
+            position: 'center',
+            icon: 'failed',
+            title: `Edit Failed`,
+            showConfirmButton: false,
+            timer: 1500
+        })
     });
 }
 
 function Delete(nik) {
-    $.ajax({
-        url: `https://localhost:44327/api/employees/${nik}`,
-        type: "DELETE"
-    }).done((result) => {
-        alert("Employee with NIK " + nik + " just got deleted");
-        location.reload();
-    }).fail((error) => {
-        alert("Failed to delete " + nik + ". Something's wrong");
-        location.reload();
-    });
+    Swal.fire({
+        title: `Do you want to delete ${nik}?`,
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            $.ajax({
+                url: `employees/${nik}`,
+                type: "DELETE"
+            }).done((result) => {
+                Swal.fire({
+                    title: "Delete Successful", text: `The employee data with NIK ${nik} has been deleted`,
+                    type: "success"
+                }).then(function () {
+                    location.reload();
+                }
+                );
+            }).fail((error) => {
+                Swal.fire({
+                    title: "Delete Failed", text: "Something went wrong!",
+                    type: "failed"
+                }).then(function () {
+                    location.reload();
+                }
+                );
+            });
+        }
+    })
 }
